17906 FIX Custom files deployed incorrectly under Windows
This Werk fixes a regression that occurs since Checkmk 2.4.0.

When deploying custom files with the ruleset _Deploy custom files with agent_, the agent bakery resolved all
_logical paths_ directly to the Windows agent installation directory.

The correct behavior is to place all files under the logical path `lib` to the installation directory, while
creating all other logical paths with files under it as placed on the Checkmk site.

Please refer to the _Deploy custom fules with agent_ ruleset help for details about _logical paths_ and
target directories in the final agent installation.

CMK-23497

Change-Id: Iebe84ba2b1fff8a2a6de41a4548d88097817b8ea