FIX 17884 Ruleset API: Fix NotificationParameter discovery
This change is relevant for plugin developers using `cmk.rulesets.v1.rule_specs.NotificationParameters`

The discovery of the `NotificationParameters` using the
```
rule_spec_<var_name> = NotificationParameters(...
```
syntax was previously broken and is with this werk now working as intended.

Note that in order to use the parameters the `name` has to match the name of the notifications script.

CMK-23388

Change-Id: I13e50e5a4fb468fdfb08d4d0dad61e4dfac88a9c