18052 FIX Piggyback Hub: Fix rule for self-monitoring of RabbitMQ
In version 2.4.0 the piggyback hub, RabbitMQ and self-monitoring for them were introduced.
Even when the hub was disabled, an automatically added "_Process discovery_" rule could pick up related processes and lead to the discovery of a "_Process * rabbitmq_" service, which would transition to CRIT after a restart or reboot.

With this werk the pattern in the rule gets adjusted to only count "rabbitmq-server" processes.

SUP-23722

Change-Id: I7c2c696af20539314009c4df4e3bf63185c7ec1e