Get rid of default global JSONEncoder patching in GUI context
We now provide a `CustomObjectJSONEncoder` which can be used in
`json.dumps` where needed.

It is a bit unclear where this actually is needed. Three potentially
relevant types were identified:

  * SessionInfo
  * LazyString
  * HTML

The to_json of the first one looks unused and for the other two it is
close to impossible to analyze all call sites since both are widely used
in the code base. Therefore we will have rely on our test chain and make
a brave try to change it and then fix the remaining issues as a
follow-up.

This removes another place where we modify a global state from a context manager.

Change-Id: I17db40fd77ea31fa2b958602983ade9e7c9002b8