18282 emerson_temp: avoid crashes when parsing
Only use the first two sensor values, as values beyond that seem to be
handled in a different structure that we lack a concrete definition for.

A more robust approach would be to come up with a good condition on
which devices are affected, but this was not picked up in this change to
not break existing setups.

CMK-23523

Closes: #818
Co-authored-by: Moritz Kiemer <moritz.kiemer@checkmk.com>
Co-authored-by: Logan Connolly <logan.connolly@checkmk.com>
Change-Id: Iccae1567786b63afe451e7e410808cf8599d1c59