18053 otel-collector: Export log messages to event console
With this werk the configuration for the Open Telemetry collector gets extended such that incoming log messages on the OTLP endpoints can optionally be forwarded to the event console.

The configuration dialog features a warning to inform users that enabling this feature can put serious load on Checkmk and potentially cause an outage.

CMK-23164 CMK-23577

Change-Id: I5741be6e3c5d1704971217ce74efd927b8268315