17865 FIX Predefined conditions: Preserve user input after input error
Creating or editing "Predefined conditions" with host/service labels set while an input error is in place - e.g. an empty "Title" field - resulted in a loss of the host/service labels input.

This behavior is fixed. When running into an input error the host and service labels input is preserved.

Change-Id: Ib6918d7c71ee58c91536a213aa49b480cbb891f2