17908 FIX Regression: Linux agent - Agent controller not working in single directory deployment.
This Werk fixes a regression that occurs since Checkmk 2.4.0p1.

When deploying the Linux agent in single directory deployment by activating the ruleset
_Customize agent package (Linux)_, the agent controller won't be able to start.<br>
This only happens in root deployment - Non-root deployment is not affected.

As a result, the agent port won't be provided and the host can't be monitored.

The agent updater is not affected, so you can fix the installation with an (automatic)
agent update.

Change-Id: I65c915f01f11547c68d4e0cbe10468f0d6116817