Enforce socket protocol in mk_mysql
With multiple instances per host the plugin may or may not show the correct data for each instance. It may happen, that all graphs show the same instance.

This depends on the mysql client tool. It has some builtin logic to decide if it prefers TCP Port connctions or socket connections.

E.g. on debian 12, mariadb-client-10.11.11 the behaviour depends on how the host is specified.
With host=localhost, the socket is used, with host=127.0.0.1 the (default) TCP port 3306 is used.

When setting up the plugin in the agent bakery, the prefilled host field reads 127.0.0.1 and thus caused me to see false data for my second instance (as both were queried at the same default TCP port).

I suggest to always use the --protocol=socket option together with the --socket=.... option. This enforces that the specified socket parameter is actually used.

The --protocol=socket option exists with all mysql mysaladmin tools since version 4.1 (and also in mariadb / mariadbadmin)