18116 bi: integrate filesystem and storage
By making use of the filesystem and storage helpers, we've now migrated
away from storing compilations and aggregations by id, and now have a
standardized way of storing the items as hashes.

In this change, we've also simplified the way that we update and fetch
metadata like the last compilation date and the latest configuration
timestamp from the filesystem.

Lastly, a migration script will run during update that will clear the
compiled aggregations and migrate the legacy frozen aggregations to the
new format.

CMK-22873

Change-Id: I1dd4dc6a67b1019eef6a28421cfaa09b0616b958