17975 FIX rule_search: allow none_of & one_of when searching for tags
The refine search when searching for tags was broken.  If the rule had
tags set using any_of or one_of, these rules would not be found. Also,
the search itself did not allow for any_of or one_of.  With this werk
we have now addressed both of these issues and you should now be able
to search rules with these conditions.

SUP-20238

Change-Id: Ifd7fcf65a812c150df0ad01f768d1765ba00c40e