bi: move redis store logic into storage module
The storage logic for the Redis cache does not need to be a part of the
compiler class. Migrating it out of the compiler module and into the
storage one, makes it easier to test. We still require the public
`is_part_of_aggregation` method as the frontend calls it, but the rest
can be factored out.

CMK-23756

Change-Id: I58fdb4b40bceca946936b82b82dfea0ec8e50ced