Build SaaS edition for Ubuntu 22.04 and Ubuntu 24.04 (again)
SaaS is running on Ubuntu 24.04 but the built docker image is based on
IMAGE_CMK_BASE which is as of now Ubuntu 22.04.
During a release the docker image build job for SaaS took way longer
than for other editions as that job had to build the Ubuntu 22.04
package as it was not built by default during the package build process.

CMK-23521

Change-Id: Iad2baf7d3a342b5c013ea706e1cb8db7034b75cb