17543 Windows Agent supports clean installation
This option was available in previous versions but
stopped working properly due to a regression introduced
in version 2.3. As a result, it was removed from the
installer's GUI in version 2.4.

Starting with this version, the issue has been resolved.
The option now works correctly again and has been
reintroduced in the installer GUI.

Change-Id: Iee208b6ae7d5ee26eec29381d4d082b5654e5077