18055 FIX update_config: Prevent error in prediction_files_cleanup due to ".info" suffix in hostname
If predictive levels were configured in a rule and a host had a ".info" suffix, an `IsADirectoryError` exception was raised during the update process.

While walking through the file system, the update config action mistook the host directory for a data file of the predictive levels feature and tried to read from it.
This is now fixed.

SUP-23872

Change-Id: I8c153bda3f35a1fe60ae3e37c64fd0e2bfc4fce6