18114 FIX Fixed Windows agent mk_inventory plugin output corruption
This werk resolves a problem with incomplete output from the Windows agent
`mk_inventory` plugin.
Previously, the `mk_inventory` plugin used the null character (`ASCII 0`)
as a field separator, which could cause the plugin execution to stop and only
partial data to be delivered.
As a fix, the separator was changed to the Unit Separator (`ASCII 31`),
which prevents this issue.

Change-Id: Iba516212f4600646bc1542b53f12cb320735fc5e