17774 jolokia_jvm_threading_pool: Improved handling of unlimited max threads and extended ruleset
The Jolokia JVM ThreadPool check now explicitly handles the case where the maximum number of threads is unlimited (`maxThreads == -1`). In this situation, the check will always return an OK state with the summary "Maximum threads: not set (unlimited)".
Additionally, the ruleset for this check has been extended. It is now possible to configure absolute threshold levels for thread counts, in addition to the previously available percentage-based thresholds.

SUP-23435

Change-Id: I804dccbe2455e49e93bf73d41514949248e42749