18123 Render DCD rules if DCD status query crashes
If there was an error thrown querying the status of the DCD on the Setup
> Dynamic host management page, the table showing the connections would
only be partly shown. This stopped you from editing the connections if a
connection rule was cause for the crash.

With this werk, we make sure that the connections are rendered and can
be edited even if there was a problem reaching the DCD.

SUP-23981

Change-Id: If359a4009bcfcaa21e468d1d8f165951c4b1a683