18119 fix: special handling for single choice tag
The deleted code is legacy code that was needed when we supported the
null tag. Now that the tag is not a valid choice, we don't need to
include this branching. When there is only one available choice, then it
will be shown in the dropdown as the default choice and the only choice.

A migration is required to edit existing rules with null tags.

SUP-24075

Change-Id: I2109bd29ee056e6d0c47fa4f2154b84bcbc0afbf