18120 wato: offer dangling folder cleanup option
When removing a site from distributed monitoring, an error could arise
saying that hosts and folders still exist associated with a given site.
The error message offered a link to the host search which makes it easy
to handle dangling hosts, but there was no help for deleting folders. In
order to clean up these dangling folders, on would needed to manually do
this via the command line.

With this change, we are now offering links to all related folders along
with a link to automatically delete these folders if said folders are
empty. The flow is constructed in such a way that a user will only see
this action if all hosts related to a site have been cleaned up and all
nested hosts and folders _unrelated to the site_ are cleaned up. That
way all folders should be empty when you are presented with this option.

Otherwise, the user must manually clean up the hosts and sub-folders
nested inside a site folder before proceeding. Use the folder links
provided by the error output to streamline this process.

SUP-20855

Change-Id: I94559d830406aae043148ea8741174a962b9094f