Simplify cmk.utils.config_path, part 8
Use *scoped* locking, which is the only sane way to do it.  Note that unlike
the other changes in this commit chain, this is *not* a refactoring, it
changes semantics: Previously the lock on serial.mk was still held after
leaving next(), and it was only unlocked later by some weird magic releasing
all held locks.  (Finding those places for all the different scenarios is
left as an "easy" exercise for the interested reader.  ;-) Now the lock is
only held during the update of the file contents, as it should be.  If some
code relies on the old weird behaviour, it should be considered a bug and
that code should use another lock for the more "global" locking it is trying
to achieve.

In other words: For every lock it should be crystal-clear *what* it is
protecting, and it should be blindingly obvious (locally!) *how long* the
lock is being held.  Everything else is highly error-prone and basically not
maintainable.

Change-Id: I15d02d85109c4b24ddefc68625daa8ccf3eb6dc0