18106 FIX OTel: Improve tmpfs usage
This change is relevant if you use the Open Telemetry integration.
When a lot of data is sent to the OTel collector endpoint of the site, the tmpfs can fill up quite fast due to temporary storage of the incoming data.
This storage is now cleaned up more often to prevent the system running out of memory.

Note: If your check interval is configured to be less then a minute, data may be lost.
Change-Id: I4f689ba78cb53f3bf714c593381001df4ae7294d