17976 FIX check-httpv2: show error source on reqwest errors in check output
Due to a version update with the dependency 'reqwest' in check-httpv2 from v0.11.27 -> v0.12.12,
the output on reqwest errors has changed. This meant that the output of the check-httpv2 check was
no longer showing the source of the error in the output.

For example, a connect error before may have looked like this in Checkmk v2.3

"error sending request for url (url): error trying to connect: invalid peer certificate: UnknownIssuer"

but in Checkmk v2.4, the source of the error was missing and looked like this.

"error sending request for url (url)"

With this werk, we now parse the sources of the error so that the end user has this information like before.
The above error will now look like

"error sending request for url (url): client error (Connect): invalid peer certificate: UnknownIssuer"

SUP-23928

Change-Id: I8ba1e025f72173f824476ca4656dadecf532a2e3