18211 FIX Deinstallation of MKP causes GUI to crash
When an MKP, that contained a dashboard or view, was uninstalled certain empty folders were deleted, which where explicitly expected to exist.
This caused all the pages in web UI of the site to crash with the message `FileNotFoundError ((2, 'No such file or directory'))`.

This could happen either because of a manual deinstallation or during an update when outdated MKPs get deactivated.

This fix ensures that these directories exist.

CMK-23624

Change-Id: I3bc292343136ae7cc1939d41da36bd5def70550b