18125 Fix naming of alert handler option in livestatus status rule
We make a product distinction between event handlers and alert handlers
(see Werk #8275). In the "Checkmk site performance and settings" rule,
there is the option to set the monitoring state when these handlers are
disabled.

This rule refered to them as "event handler", which in the enterprise
edition should read "alert handler". Now the name lists both names to
avoid confusion.

BETA-107

Change-Id: Ifd5121646defde566ffa7f35ce00de1484010e78