17754 FIX Do not open an unlimited amount of file descriptors in the EC
Previously, the event console accepted an unlimited amount of concurrent
connections via TCP and the UNIX domain socket.  This could lead to running
out of file descriptors in the EC process.  We limit the number of
concurrent connections now to a fixed value to avoid this. When reaching
that limit, any new connections are queued in the OS queue, and when even
that is full, the client gets an error, which is the right thing here.

Note: The limit is currently at a relatively arbitrary high number (900
concurrent connections), making this configurable is probably not needed.
The only essential thing is that we have some breathing room below the hard
limit of 1024 file descriptors caused by the usage of select().

Change-Id: Ie3ad0a4b05a922d179397ca4ab6ec5c5c127a45e