17755 FIX Ignore empty CMC state files
When the state file of the CMC ends up being empty due to e.g. some problems in
the filesystem, it was still considered to be a valid state file. While this is
technically correct (everything is at its default then), this is not what one
wants: All switches in the master control would be "off", so monitoring would
effectively be disabled then.

We now ignore such an empty state file, so the CMC state effectively starts from
scratch, then with no acknowledgements, comments, ad hoc downtimes, etc. While
this is not perfect, it is better than a running but effectively disabled
monitoring site.

Change-Id: If339f3c1ce7949e3b22b053932148755324861cc