BI: Allow detection if query supports JSON
The raw livestatus response is only parsed as JSON if the query
supports JSON format, which can only be determined if the query is
specified using cmk.livestatus_client.QuerySpecification.
Since the Single-/MultiSiteConnection query automatically sets the
output format to JSON if it is supported, BI queries will now switch
to JSON if possible.

CMK-23079

Change-Id: I911daaf5da601dff98d14e461775327e66a91a11