Stop our buttons from submitting forms by default
We often use our vue apps inside a form context. As this is
very intransparent when you are inside of vue, we opted to
always .preventDefault the click event in our buttons so that
you don't have to keep this possibility in mind.

Change-Id: If129471e8b9126be81ec330550611bacbcc849a4
(cherry picked from commit 4ca86b776446e2342dc07fae7d544d903edfcf25)