Implement frontend internationalization stub
We will ship 2.5.0 with frontend i18n. However, features currently in
development should not be concerned with feeding the i18n via the
backend.

Therefore we implement a t() i18n stub interface to have a single
entry point which we can later replace with a proper key-based
i18n implementation.

Change-Id: I3be21743eb91f12d1c1d09d2727e39c7f90681b7