18249 FIX OTel: Fix metric name for OTel datapoints without attributes
With [werk #18209](https://checkmk.com/werk/18209) and [werk #18108](https://checkmk.com/werk/18108) we changed the way the Checkmk metric name is being computed from the OTel data points by using the attributes of the data points and the unit of the OTel metric.
If the OTel metric contains only a single data point, the data point usually has no attributes, resulting in no Checkmk metric being emitted in the worst case.

For datapoints without attributes we now use the name of the OTel metric instead of the attributes to build the Checkmk metric name.

Please note that because of this change, metrics which were previously created by the OpenTelemetry integration, will no longer be supplied with data. Instead, new metrics following this new naming convention will be created.

CMK-24072

Change-Id: I353e9e1e7f18e6cf6fc1aaaf1040b6d477ca0a7e