k8s PoC: Build package with k8s
The used pod template is extended in the backend code with the
specified distro and docker tag.

Unless the "kubernetes_inherit_from" variable is set to some pod template
the existing way of container usage is used. Making it fully and finely
granular backwards compatible.

Requires https://review.lan.tribe29.com/c/checkmk_ci/+/105044

Only ugliness is the signing of the packages which has to be performed
in a nother container based on Ubuntu 22.04, see CMK-24094

copyArtifacts seems to be incompatible with k8s and their containers.

CMK-23959 and CMK-24355

Change-Id: I4ee875f883a97b693e2fbed3645751b48b64cd78