17977 user: you can now select the start_url for a user via the Rest-API
Due to some recent updates, we now have three options for
selecting a default start up page for each user. Previously,
you could set it to a default dashboard or any other custom
dashboard. We now have a third option, which is a welcome page
for new users.

This change allows the rest-api user to select any one of
these three options when creating or editing a user. The
default setting will be the default dashboard.

CMK-24133

Change-Id: Ife9e84a368ba106754887cdd646b0c7622d5c80f