Specify a relay protocol for monitoring data
Monitoring data is send to the site on regular intervals
so we now have a defined schema for it. The implementation
is using pydantic to define the schema

The site mock is updated to setup a stub on the data endpoint
with the schema we defined. This ensures anything we send to the
stub confirms to our schema. We are using the pydantic definition
directly to avoid code duplication

Only little annoying thing is when constructing the MonitoringData
model we immediately need to base64 encode the payload.

Change-Id: I2eb24ccad368d701a97b2e32afeecbfe3d4064d4
JIRA-Ref: SAASDEV-4558