17978 FIX ldap: align start_url user attribute with user_config
We now have three options for a user to set their
start_url. These are "default_start_url", "welcome_page"
or any other custom string.

This werk makes sure that the user_attribute to sync
available for an ldap connection, aligns with the
same values available for the user_config.

Change-Id: Id827562449312529a0312622bbe2ae6b11d51bfb