17909 FIX Agent Bakery: TMPDIR environment variable was not applied for single directory setup
When setting the bakery rule _Customize agent package_, you can optionally set/overwrite
the `TMPDIR` environment variable in the scope of the agent.

Previously, this settng was not applied correctly, leaving the `TMPDIR` variable untouched
on the target system.

This Werk is marked as incompatible because you have to update/reinstall the Checkmk agent
on affected systems to apply the fix.

Change-Id: Id50069d9aa350f3f92cdc9b670cc7e0bc19822c1