18213 FIX Prometheus special agent: Circumvent maximum length of argument list
For very extensive rulesets for the Prometheus special agent the passed arguments could exceed the system limit for the argument list.
This resulted in a `OSError(7, 'Argument list too long')`, which prevented the special agent from running.

This werk passes the more extensive parts of the data via a pipe to STDIN of the special agent instead of passing them as CLI arguments.

SUP-24239

Change-Id: I175d7836136d0926ca9285a5c38b58ff7edf2652