wa: remove flaky test in mk_mysql.Tests.ps1
The test was executing a real system cmd.exe call,
which is not appropriate for unit testing. It was unstable
and sometimes passed or failed. Pester does not perform well
with external calls, so such calls should be mocked, but in
this case mocking does not make sense. The test has been removed.

Change-Id: Ib10439e8cb949626139ac2c3c15c32bd13969095