15367 FIX passwordstore special characters
Allow special characters in passwords saved in the passwordstore, read
by a c-compiled active check

The c-program that replaces the `*` with the actual password is first
checking if there is enough room for the real password. It's not if the
real password contains special characters that are encoded into multiple
bytes.

This can be easily reproduced by create an entry in the password store
with an single special character password (`ä`) and using that password
in the authentication of check_http active check.

You can extract the active check command line by running
  cmk -N | grep http
you will probably see something like:
  check_command                 check_mk_active-http!--pwstore=2@10@a -a 'something:*' --sni -I localhost
this will be executed as:
  ~/lib/nagios/plugins/check_http --pwstore=2@10@a -a 'something:*' --sni -I localhost
and will return an error:
  Invalid --pwstore= option: Password is too long for argument.

You can then manually fix that by replacing that single `*` by `**`.

This change encodes the password using utf-8 before calculating the
length.

SUP-24322

also reformatted tests/unit/cmk/base/test_server_side_calls.py

Change-Id: Ia6383c1efa64ea4a2ec0456d150b62e58b25954d