Pass UI Config object to page handling functions
This hands over the `Config` object which is so far mostly accessed using
the `active_config` global proxy to the page handling functions so that
we can use the passed config object instead of the global proxy moving
forward.

Passing the `Config` to `Page.page` is still missing and will come in a
follow-up change.

Change-Id: Ib308fd534c6e3a7951fc2ff4f8ad3c01fa33df48