ip lookup: make family mandatory
While we can determine the default address family
from the same config (currently), it does make sense
from a semantical point of view to make this parameter
mandatory for a function that looks up an IP address.

This will also allow us to pass the pre-made lookup
function to the `ConfiguredIPLookup` class,
which in turn will enable us to get rid of the global
states (or at least ban them into the modes module).

Change-Id: I22fbda76aa9a97c7fd0709dd12ccfd551247c94a