Add Fetcher class
This class is used to start a fetcher process.
You can schedule a task on it, and any result
it placed in a main queue.

Since we handle subprocess behavior, the fake fetchers
are really useful here. Maybe the testlib should later
be used by unit and component tests. lets see if we have
more use for it.

Change-Id: I8a9135c2686ec368f25f1d4afdaeb554ba78a037