18473 agent_ipmi_sensors: Add option to ignore unrecognized events
Add new option "Ignore unrecognized events" to the "Use FreeIPMI"
section of the "Agents" rule "IPMI Sensors via Freeipmi or IPMItool",
which will add the `--ignore_unrecognized_events` argument to the
`ipmi-sensors` invocation.

With this new option it is possible to ignore unrecognized events which
would otherwise result in a critical service state, as well as entries
like "Unrecognized Event = 0100h" in the service summary.

CMK-24393

closes #833

Change-Id: I8f8c3bcf94a7fbc8b03e3cdabfff5f4c66524ba1