Split FrontendModel from InvalidValue model
We gain nothing from making the return of _to_vue anything else than
object. There is no typed consumer of this data. We lose however by
linking the InvalidValue fallback_value to the same as the return of
_to_vue, since we sometimes need to transmit data to _to_vue inside
of the InvalidValue which has nothing to do with the frontend (see
dictionary).

Change-Id: I70de62f525d97e54e561d668fdc2d7a035000d89