17979 FIX site management: users attempting to login to a remote site do not have to exist on the main site
In the UI, via distributed monitoring, it is possible to login to a remote
site with a user that does not exist on the main site. However, this was not
possible via the Rest-API, as we were checking that the username provided
existed on the main site.

With this werk we have changed this behaviour to match that of the UI. The
username provided to the Rest-API login endpoint no longer needs to exist
on the main site.

SUP-23881

Change-Id: I0bc9d895e9ceec9a002beb6fecb539277f8b708d