Remove DataOrigin in form spec architecture
We receive two kinds of data formats in the form spec visitors: data
that is coming from disk and data that is coming from the frontend.

The way our public API was structured allowed the visitors to expect
disk data but be passed frontend data. This was the source of many bugs.

In this commit, we wrap the data itself to let the visitors know what
they're dealing with and remove the visitor option to specify the data
origin.

Change-Id: I12012e653da40e9994c345b553b67c99c1e97189