18169 FIX special_agent storage/proxmox_ve: url quote key and hostnames instead of checking for allowed chars
This werk fixes a bug in the proxmox_ve plugin, where the new special agent storage was used.
This change was introduced with werk [Werk #18164](https://checkmk.com/werk/18164).

The implementation used an illegal key which raises an unhandled ValueError.
The sanitation was changed from allowed chars to url quoted keys and hostnames.

Change-Id: If09fd5a8ea3fe0ba027caefa28a4f0e0ebb54513