FIX 18176 disktat: fix rulespec migration when using the API
When using the API to create rules, the diskstat rulespec would migrate provided conifguration values
and eventually drop unknown keys, as well as add missing keys without feedback.

With this werk, the migration will only add default keys and values when they are missing.
Other keys and values will be kept and will be subsequently rejected by the validation.

SUP-24017

Change-Id: Ia966aa4f5990fa2c21656c4a7552675b3ff29a3d