search: add unified search endpoint
In order to show all results in the new unified search menu, we need to
provide the frontend with a better payload structure. The current Setup
and Monitoring searches group by topic, but this is not desired for the
new search because it makes displaying this information inflexible.

Key features of the response payload from the new endpoint:

  - a new "provider" attribute was introduced, i.e. "setup".
  - a new "topic" attribute was introduced, no longer grouped.
  - meta information: result counts (by provider), url, and query.
  - results are sorted alphabetically by "title".

The endpoint, in addition to the `q` (query) parameter, also now accepts
a `provider` parameter to filter on just setup or monitoring results. If
not included, all results will be returned. Here is a URI with an
example query:

  GET /check_mk/ajax_search.py?q=host&provider=setup&_ajaxid=0

NOTE: to deliver this as soon as possible, we are making use of the
existing search engines and transforming the payload into the shape that
the frontend needs. This is of course not performant, but future
refactorings are planned to get rid of this hacky solution. The focus
now is to respond with the proper response shape.

CMK-24382

Change-Id: I76f38b3c7584871ac94f5d5c5b55e3198c8bdbbe