18294 FIX Logwatch: fix for service summary on reclassification to state OK
Currently, if the logwatch reclassify parameter is set to reclassify everything to OK, the result summary always shows "No error messages". This happens because of the reclassification to OK, even when there have been error messages.

With this werk, the summary and details will no longer be overwritten, even if all errors are reclassified to state OK. This makes it possible to see error and warning messages in the service summary, even when reclassified as OK.

Change-Id: Ic8e788e16364de9b94077c650897a1dfc835b02d