search: bring search logic together into package
The monitoring search logic does not belong in the snapin module. Let's
move it and the unified search to the watolib alongside the setup search
logic which always resides there. By converting the the `search.py`
module into a package, we don't need to update all the call site
imports.

Change-Id: I205f067567068a12f23b06e217c9283983c8916e