Enable taplo formatting over bazel-lint
We contributed and upstreamed taplo formatting to rules lint in
  https://github.com/aspect-build/rules_lint/pull/576

So let's use it in the repo!

This patches improves some naming in the `BUILD` file and fixes a bug
where the formatter would fail unless run from the repo root.

We let the `run` scripts format our `.toml` files with rules lint.

Change-Id: Ifcf734ab7f2442b6ad42b8ada2fde1e4c3ed055a