config: isolate CMCConfig
The remainig methods of the CEEConfigCache are in fact
a configuration purely needed for the microcore config
generation.

We pass it during the MonitoringCore instance creation,
to avoid passing unneeded arguments to the nagios cores
create_config method (and having to keep CEE specific
code in the raw edition just to annotate an unused
argument).

We add a few more wrapper classes to the ruleset module:
These are used to combine the ruleset matcher, the label
manager and a ruleset to a callback that can be passed
to a component that is agnostic of the implementation.

Change-Id: If898113e5fa34e665423494a194d36dc2b6fc56d