18340 FIX mbg_lantime_ng_refclock: Always create service for refclocks
When new clock types were added to the MIB definition, Checkmk was unable to determine from the ID whether it was a GPS or non-GPS clock and did not discover it as a service.

With this change, clocks will always be discovered as a service.
If in doubt, it will be assumed that it is a non-GPS clock.
The service state for unknown clock types will always be WARN or worse.

SUP-24487

Change-Id: I6d000ad7b68d1df992128677b75cdf0bad3348a8