18170 otel: Allow metrics of different instances of the same resource to be aggregated
When using the otel collector to collect resource metrics from a application that has multiple replicas*,
currently only the last metric that has reached the collector will be used for the check,
independent of replicas.

With this werk different replicas with the same metrics can be aggregated (min, max, sum, average).
This way the last received metric of each replica will be used for the value of the check.

The default value will keep the old behavior and will use the last received metric independent of replica.

*multiple replicas in this context means resource metrics that have different resource attributes beyond
what is used for host name computation.

CMK-24076

Change-Id: I2548581f07d4b30cb6ba4cc7a66d5e59c36ff54f