Avoid mutation of a list while iterating over it. Some cleanups.
Although we are lucky here, pop()ing from a list while enumerate()ing
over it is bad style and leads to a bug more often than not.

We avoid this by using a functional approach, and we use this at other
places, too.  This enables us to have a clean and consistent pattern for
loading => transforming => saving messages.

Change-Id: Ib07895c00f9596fc8a17491527904a3f091ed43a