18341 FIX snmplib: Always use SNMPv3 contexts in requests
When an SNMPv3 device (e.g. a printer from HP or Brother) required a context, the connection test and service discovery failed with the following error:

```
SNMP Error on <hostname> while walking .1.3.6.1.2.1.1.1. Normally this is caused by a device sending invalid SNMP responses (Details: Bad context specified (0/-17)).
```

This error was caused by an incorrect assumption in the filtering of the "SNMPv3 contexts to use in requests" rules.

With this change, the filter algorithm returns the context valid for all SNMP sections even if the request was not specific to a single SNMP section.

SUP-24342

Change-Id: I332c17492ce4c3101a43a3b2b057bf4bf6c7fd86