otel: limit file ingestion of special agent
When the cleanup job wasn't running, the special agent has been high number of files and therefore data.
This would lead to a high RAM usage, a unresponsive site and subsequently to a system crash.

To avoid overusage of system resources and crashes of the checkmk site,
the amount of files that the special agent reads has been limited to files
that have been written within the last 65 seconds.

CMK-24598

Change-Id: I4b9f636c0b5a95bf0ea4dfd71db5955dd346621a