Fix import cycle between legacy converter and wato
Looks like this cycle was revealed by the reccently changed import
order. While the internal jobs were not affected, the github actions
broke, can be reproduced with `pytest tests/unit/cmk/gui`.

To break this cycle, this change moves the dependencies of the legacy
converter (rulespecs) to cmk.gui.watolib.

Change-Id: Ib0a92cad14049f3bd32487f826b09334d8211bba