17934 Added temperature support for Aruba 6100 switch series
SUP-24444

For the Aruba 6100 switch series, no temperature was discovered and
checked.

The support for this device series now has been added.

The device defines the thresholds for low and high temperature by itself
and jumps immediatly from OK to CRIT.
When sensor is in faults state, this check is in WARN state.

The check uses the arubaWiredTempSensorEntry-MIB.

Change-Id: Ied80fff5d68ad6c6399aa630778bcc8d9f8faa92